Restructure dialog
